% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mli0}
\alias{mli0}
\title{mli0}
\format{A SpatialPolygonsDataFrame with 1 feature:
\describe{
\item{\code{OBJECTID}}{Object identifier}
\item{\code{admin0Name}}{Name of administrative level 0 (country)}
\item{\code{admin0Pcod}}{Postal code of administrative level 0 (country)}
\item{\code{admin0RefN}}{Reference name of administrative level 0 (country)}
\item{\code{admin0AltN}}{Alternative name of administrative level 0 (country)}
\item{\code{admin0Al_1}}{Alternative name of administrative level 0 (country)}
\item{\code{date}}{Date map was poduced}
\item{\code{validOn}}{Date map was last validated}
\item{\code{validTo}}{Date map is valid up to}
\item{\code{Shape_Leng}}{Lenght of shape}
\item{\code{Shape_Area}}{Shape of area}
}}
\source{
Humanitarian Data Exchange <https://data.humdata.org/dataset/administrative-boundaries-cod-mli>
}
\usage{
mli0
}
\description{
Administrative Level 0 (Country) map of Mali
}
\keyword{datasets}
