% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mli2}
\alias{mli2}
\title{mli2}
\format{A SpatialPolygonsDataFrame with 50 features:
\describe{
\item{\code{OBJECTID}}{Object identifier}
\item{\code{admin2Name}}{Name of administrative level 2 (cercle)}
\item{\code{admin2Pcode}}{Postal code of administrative level 2 (cercle)}
\item{\code{admin2RefN}}{Reference name of administrative level 2 (cercle)}
\item{\code{admin2AltN}}{Alternative name of administrative level 2 (cercle)}
\item{\code{admin2Al_1}}{Alternative name of administrative level 2 (cercle)}
\item{\code{admin1Name}}{Name of administrative level 1 (region)}
\item{\code{admin1Pcod}}{Postal code of administrative level 1 (region)}
\item{\code{admin0Name}}{Country (administrative level 0) name}
\item{\code{admin0Pcod}}{Country (administrative level 1) code}
\item{\code{date}}{Date map was poduced}
\item{\code{validOn}}{Date map was last validated}
\item{\code{validTo}}{Date map is valid up to}
\item{\code{Shape_Leng}}{Length of shape}
\item{\code{Shape_Area}}{Area of shape}
}}
\source{
Humanitarian Data Exchange <https://data.humdata.org/dataset/administrative-boundaries-cod-mli>
}
\usage{
mli2
}
\description{
Administrative Level 2 (Cercle) map of Mali
}
\keyword{datasets}
