% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mli3}
\alias{mli3}
\title{mli3}
\format{A SpatialPolygonsDataFrame with 701 features:
\describe{
\item{\code{OBJECTID_1}}{Object identifier}
\item{\code{admin0Name}}{Country (administrative level 0) name}
\item{\code{admin0Pcod}}{Country (administrative level 1) code}
\item{\code{admin1Name}}{Name of administrative level 1 (region)}
\item{\code{admin1Pcod}}{Postal code of administrative level 1 (region)}
\item{\code{admin2Name}}{Name of administrative level 2 (cercle)}
\item{\code{admin2Pcod}}{Postal code of administrative level 2 (cercle)}
\item{\code{admin3Name}}{Name of administrative level 3 (commune)}
\item{\code{admin3Pcod}}{Postal code of administrative level 3 (commune)}
\item{\code{admin3RefN}}{Reference name of administrative level 3 (commune)}
\item{\code{admin3AltN}}{Alternative name of administrative level 3 (commune)}
\item{\code{admin3Al_1}}{Alternative name of administrative level 3 (commune)}
\item{\code{date}}{Date map was poduced}
\item{\code{validOn}}{Date map was last validated}
\item{\code{validTo}}{Date map is valid up to}
\item{\code{Shape_Leng}}{Length of shape}
\item{\code{Shape_Area}}{Area of shape}
}}
\source{
Humanitarian Data Exchange <https://data.humdata.org/dataset/administrative-boundaries-cod-mli>
}
\usage{
mli3
}
\description{
Administrative Level 3 (Commune) map of Mali
}
\keyword{datasets}
